Update diff files which break benchmark suite
Several changes occurred for specific projects since last update, which
caused building failure for the benchmark suite. This patch fixes all of
them.

1) Panorama project becomes warning-free, so I fixed all the warnings
in the patch for panorama benchmark.

2) Skia and dex2oat projects have modifications on certain lines in the
bp file, so I updated the patch.

3) Binder no longer needs to include the test directory.

4) Skia becomes static linked to benchmark, so libskia no longer exists.
Modified autotest patch so that we do not try to upload libskia.so to
device any more.

Test: Rebuilt all the benchmarks in the suite and they all work fine
now.

Change-Id: I9a28bd999e197acae1e8ca434f3db68e1f4dcda9